Support MSYS and MinGW without MSYS

+++ GNUmakefile
@@ -4,18 +4,33 @@ BINARIES = bin
 LIBRARIES = lib
 BUILD = build
 
-ifeq ($(findstring MINGW, $(shell uname -s)), MINGW)
+UNAME = $(strip $(shell uname))
+
+ifeq ($(UNAME),)
 	GENERATOR = "MinGW Makefiles"
+	IS_MINGW = 1
 else
-	GENERATOR = "Unix Makefiles"
+	ifeq ($(findstring MINGW,$(UNAME)),MINGW)
+		ifeq ($(findstring msys,$(shell sh --version)),msys)
+			GENERATOR = "MSYS Makefiles"
+			# Workaround for strange path bug.
+			MAKE_COMMAND = /usr/bin/env make
+		endif
+	else
+		GENERATOR = "Unix Makefiles"
+	endif
 endif
 
 all: | $(BUILD)
 	cd $(BUILD) && cmake -G $(GENERATOR) ..
-	$(MAKE) -C $(BUILD)
+	$(MAKE) --no-print-directory -C $(BUILD)
 
 clean:
+ifdef IS_MINGW
+	RD /S /Q $(BUILD) $(BINARIES) $(LIBRARIES) 2>NUL || REM
+else
 	$(RM) -r $(BUILD) $(BINARIES) $(LIBRARIES)
+endif
 
 test:
 	cd $(BUILD) && ctest
